added service scripts

+++ datalogger/telemetry.py
@@ -34,7 +34,6 @@ MQTT.configureOfflinePublishQueueing(100000, AWSIoTPythonSDK.MQTTLib.DROP_OLDEST
 MQTT.configureDrainingFrequency(25)  # Draining: 25 Hz
 MQTT.configureConnectDisconnectTimeout(10)  # 10 sec
 MQTT.configureMQTTOperationTimeout(5)  # 5 sec
-MQTT.connect()
 
 def closeConnections():
 	try:
@@ -45,43 +44,47 @@ def closeConnections():
 		pass
 
 while True:
-	try:		
-		## DATABASE CONNECTION ##
-		DB = sqlite3.connect('sensordata.db')
-		cursor = DB.cursor()
-		cursor.execute('CREATE TABLE IF NOT EXISTS sensor (sensor INT NOT NULL, timestamp INT NOT NULL, payload CHAR(500) NOT NULL);')
-		DB.commit()
-		try:
+	try:
+		if os.path.isfile('certs/root-CA.crt') and os.path.isfile('certs/private.pem.key') and os.path.isfile('certs/certificate.pem.crt'):
+			## mqtt CONNECTION ##
+			MQTT.connect()
+
+			## DATABASE CONNECTION ##
+			DB = sqlite3.connect('sensordata.db')
 			cursor = DB.cursor()
-			cursor.execute('CREATE INDEX timestamp ON sensor (timestamp);')
+			cursor.execute('CREATE TABLE IF NOT EXISTS sensor (sensor INT NOT NULL, timestamp INT NOT NULL, payload CHAR(500) NOT NULL);')
 			DB.commit()
-		except:
-			pass
-		logger.info('DB CONNECTION OK.')
-
-		while True:
-			q = conn.execute('SELECT * FROM sensor ORDER BY timestamp LIMIT 100;')
-			to_timestamp = 0 # epoch
-			from_timestamp = int(time.time())+315360000
-			
-			for r in q:
-				if r[1] > to_timestamp:
-					to_timestamp = r[1]
-				if from_timestamp > r[1]:
-					from_timestamp = r[1]
-
-				payload = {
-					'devid': DEVID,
-					'sensor': r[0],
-					'timestamp': r[1],
-					'payload': json.loads(r[2])
-				}
-				MQTT.publish(CHANNEL, json.dumps(payload), 1)
-
-			if to_timestamp and from_timestamp:
+			try:
 				cursor = DB.cursor()
-				cursor.execute('DELETE FROM sensor WHERE timestamp>='+str(from_timestamp)+'AND timestamp<='+str(to_timestamp)+';')
+				cursor.execute('CREATE INDEX timestamp ON sensor (timestamp);')
 				DB.commit()
+			except:
+				pass
+			logger.info('DB CONNECTION OK.')
+
+			while True:
+				q = conn.execute('SELECT * FROM sensor ORDER BY timestamp LIMIT 100;')
+				to_timestamp = 0 # epoch
+				from_timestamp = int(time.time())+315360000
+				
+				for r in q:
+					if r[1] > to_timestamp:
+						to_timestamp = r[1]
+					if from_timestamp > r[1]:
+						from_timestamp = r[1]
+
+					payload = {
+						'devid': DEVID,
+						'sensor': r[0],
+						'timestamp': r[1],
+						'payload': json.loads(r[2])
+					}
+					MQTT.publish(CHANNEL, json.dumps(payload), 1)
+
+				if to_timestamp and from_timestamp:
+					cursor = DB.cursor()
+					cursor.execute('DELETE FROM sensor WHERE timestamp>='+str(from_timestamp)+'AND timestamp<='+str(to_timestamp)+';')
+					DB.commit()
 	except Exception,e:
 		logger.error('error', exc_info=True)
 		time.sleep(1)
